requirements-dev.txt
Django==5.1.5
django-ckeditor-5==0.2.15
pillow==11.1.0
django-imagekit==5.0.0
python-decouple==3.8
django-robots==6.1
django-rosetta==0.10.1
django-parler==2.3
django-filer==3.3.1
django-sass-processor==1.4.1
libsass==0.23.0
django-browser-reload==1.17.0
django-mptt==0.16.0
